+++ plugins/postcss-cascade-layers/.tape.mjs
@@ -43,7 +43,7 @@ postcssTape(plugin)({
 			onConditionalRulesChangingLayerOrder: 'warn',
 			onImportLayerRule: 'warn'
 		},
-		warnings: 3,
+		warnings: 4,
 	},
 	'warnings:with-postcss-import': {
 		message: 'correctly handles warnings when postcss-import is used',
@@ -54,7 +54,11 @@ postcssTape(plugin)({
 		},
 		warnings: 2,
 		plugins: [
-			postcssImport(), /* postcss-import must run first */
+			postcssImport({
+				nameLayer: (index) => {
+					return `anon-layer-import--${index}`;
+				}
+			}), /* postcss-import must run first */
 			plugin(),
 		]
 	},