+++ plugins/postcss-focus-within/src/browser-global.js
@@ -1,3 +1,6 @@
-/* global self */
+/* global self,window */
 import { default as focusWithinInit } from './browser';
-self.focusWithinInit = focusWithinInit;
+
+(function (global) {
+	global.focusWithinInit = focusWithinInit;
+}('object' === typeof window && window || 'object' === typeof self && self || {}));